pyproject.toml
[project]
name = "sopy-times"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
]

[dependency-groups]
dev = []
